fix: atomize otp failure counting
