+++ src/test/kotlin/app/xquare/xquareinfra/application/accountRecovery/AccountRecoveryServiceTest.kt
@@ -7,15 +7,14 @@ import app.xquare.xquareinfra.application.accountRecovery.ports.inbound.VerifyPa
 import app.xquare.xquareinfra.application.accountRecovery.ports.inbound.VerifyUsernameFindOtpCommand
 import app.xquare.xquareinfra.application.accountRecovery.ports.outbound.UserPersistenceForAccountRecoveryPort
 import app.xquare.xquareinfra.application.auth.AuthException
-import app.xquare.xquareinfra.application.auth.ports.outbound.PasswordEncoderPort
 import app.xquare.xquareinfra.application.emailOtp.EmailOtpProperties
 import app.xquare.xquareinfra.application.emailOtp.EmailOtpPurpose
 import app.xquare.xquareinfra.application.emailOtp.EmailOtpService
-import app.xquare.xquareinfra.application.emailOtp.ports.outbound.EmailOtpPort
-import app.xquare.xquareinfra.application.emailOtp.ports.outbound.EmailSendPort
-import app.xquare.xquareinfra.application.emailOtp.ports.outbound.OtpConsumeResult
 import app.xquare.xquareinfra.domain.user.User
 import app.xquare.xquareinfra.domain.user.UserRole
+import app.xquare.xquareinfra.testFixtures.FakeEmailOtpPort
+import app.xquare.xquareinfra.testFixtures.FakeEmailSendPort
+import app.xquare.xquareinfra.testFixtures.FakePasswordEncoderPort
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
@@ -282,124 +281,4 @@ class AccountRecoveryServiceTest {
                     it.email.equals(email, ignoreCase = true)
             }
     }
-
-    private class FakePasswordEncoderPort : PasswordEncoderPort {
-        override fun encode(password: String): String = "encoded:$password"
-
-        override fun matches(
-            password: String,
-            encodedPassword: String,
-        ): Boolean = encodedPassword == "encoded:$password"
-    }
-
-    private class FakeEmailSendPort : EmailSendPort {
-        val sentEmails = mutableListOf<SentEmail>()
-
-        override fun send(
-            to: String,
-            subject: String,
-            body: String,
-        ) {
-            sentEmails += SentEmail(to = to, subject = subject)
-        }
-
-        override fun sendWithTemplate(
-            to: String,
-            subject: String,
-            templateName: String,
-            variables: Map<String, Any>,
-        ) {
-            sentEmails += SentEmail(to = to, subject = subject)
-        }
-    }
-
-    private class FakeEmailOtpPort : EmailOtpPort {
-        private val otps = mutableMapOf<Pair<EmailOtpPurpose, String>, String>()
-        private val otpFailures = mutableMapOf<Pair<EmailOtpPurpose, String>, Int>()
-        private val verifiedTokens = mutableMapOf<Pair<EmailOtpPurpose, String>, String>()
-
-        override fun saveOtp(
-            purpose: EmailOtpPurpose,
-            email: String,
-            otp: String,
-            ttlSeconds: Long,
-        ) {
-            otps[purpose to email] = otp
-            otpFailures.remove(purpose to email)
-        }
-
-        override fun getOtp(
-            purpose: EmailOtpPurpose,
-            email: String,
-        ): String? = otps[purpose to email]
-
-        override fun consumeOtp(
-            purpose: EmailOtpPurpose,
-            email: String,
-            otp: String,
-        ): OtpConsumeResult {
-            val key = purpose to email
-            val savedOtp = otps[key] ?: return OtpConsumeResult.NOT_FOUND
-            if (savedOtp != otp) {
-                return OtpConsumeResult.MISMATCH
-            }
-            otps.remove(key)
-            otpFailures.remove(key)
-            return OtpConsumeResult.CONSUMED
-        }
-
-        override fun recordOtpFailure(
-            purpose: EmailOtpPurpose,
-            email: String,
-            ttlSeconds: Long,
-            maxFailures: Int,
-        ) {
-            val key = purpose to email
-            val failures = (otpFailures[key] ?: 0) + 1
-            if (failures >= maxFailures) {
-                otps.remove(key)
-                otpFailures.remove(key)
-                return
-            }
-            otpFailures[key] = failures
-        }
-
-        override fun saveVerifiedToken(
-            purpose: EmailOtpPurpose,
-            token: String,
-            email: String,
-            ttlSeconds: Long,
-        ) {
-            verifiedTokens[purpose to token] = email
-        }
-
-        override fun consumeVerifiedToken(
-            purpose: EmailOtpPurpose,
-            token: String,
-            expectedEmail: String?,
-        ): String? {
-            val key = purpose to token
-            val email = verifiedTokens[key] ?: return null
-            if (expectedEmail != null && email != expectedEmail) {
-                return null
-            }
-            verifiedTokens.remove(key)
-            return email
-        }
-
-        fun hasOtp(
-            purpose: EmailOtpPurpose,
-            email: String,
-        ): Boolean = otps.containsKey(purpose to email)
-
-        fun hasVerifiedToken(
-            purpose: EmailOtpPurpose,
-            token: String,
-        ): Boolean = verifiedTokens.containsKey(purpose to token)
-    }
-
-    private data class SentEmail(
-        val to: String,
-        val subject: String,
-    )
 }
